Better OAI error handling and recovery.
Introduced a notion of 'initial harvest', which, if completed without
errors, causes the harvester to ignore subsequent "noRecordsMatch"
OAI errors (raised both when no new records where added to the repo (expected)
or when some of the repository configuration elements are mis-specified,
yes OAI-PMH is very advanced). noRecordsMatch on initial harvests are treated
as configuration errors.

Also, all errors are now fully persistent and will not be cleared when
the harvester is restarted.