Abandons binding of select items
Doesn't seem to play well with other updates of the select list happening.

Was used as a hook for regenerating the list anew after a change of source.
Without that the UI control and the source list had to be hooked up by
other means - now inserting a source ID in the page in a composite element
and looking up that element via the change event to fetch the source ID.