Use manual XmlAdapter rather than @XmlIDREF
Fixes mysql constraint exception blocking resolution of MKH-439

TL;DR

The exception was caused by 'transformation' member being set to null
when unmarshalling TSA entity from XML.

The 'transformation' member was represented in the XML via the @XmlIDREF
annotation that does not serialize the whole entity but only its ID,
this was done to avoid circular references between Transformation and TSA.
@XmlIDREF does not look up the member entity during unmarshalling, though,
and leaves the member set to null. This in turn will trigger NOT NULL constrain
violation when persisting the entity.

The solution is to write a manual (un)marshalling adapter via XmlAdapter
that creates a dummy instance of Transformation and sets the proper ID on it.
The actual instance is not looked up in the persistence context, but it
does not seem to cause problems.

Other solutions, like using the @XmlInverseReference annotation where tried
but did not work.

With this commit the default JAXB adapter is changed to MOXy (provided
with EclipseLink) which seem to be more aligned with this particular JPA
provider.